Refactored pages into their own directory
