+++ frontend/src/App.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import { useAuth0 } from '@auth0/auth0-react';
-// pages
-import Landing from './Landing';
-import Tracks from './Tracks';
-import Setlist from './Setlist';
-import Home from './Home';
-import Artist from './Artist';
-import TourDash from './TourDashboard';
-// components
-import { SideBar, Loading, TopBar, PrivateRoute } from './components/index';
-// styles
 import './styles/main.css';
+
+import {
+  Landing,
+  Tracks,
+  Setlist,
+  Home,
+  Artist,
+  TourDashboard,
+} from './pages/index';
+import { SideBar, Loading, TopBar, PrivateRoute } from './components/index';
 
 export default function App() {
   const { isLoading, isAuthenticated } = useAuth0();
@@ -36,7 +36,7 @@ export default function App() {
         <PrivateRoute path="/tracks" component={Tracks} />
         <PrivateRoute path="/artist" component={Artist} />
         <PrivateRoute path="/setlist" component={Setlist} />
-        <PrivateRoute path="/tourdashboard" component={TourDash} />
+        <PrivateRoute path="/tourdashboard" component={TourDashboard} />
       </Switch>
     </Router>
   );
