+++ frontend/src/pages/Home.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useAuth0 } from '@auth0/auth0-react';
-import { DashboardTile, ActivityPanel } from './components/index';
+import { DashboardTile, ActivityPanel } from '../components/index';
 
 export default function Home() {
   const { user } = useAuth0();
