+++ frontend/src/pages/index.tsx
@@ -1,0 +1,8 @@
+import Artist from './Artist';
+import Home from './Home';
+import Landing from './Landing';
+import Setlist from './Setlist';
+import TourDashboard from './TourDashboard';
+import Tracks from './Tracks';
+
+export { Artist, Home, Landing, Setlist, TourDashboard, Tracks };
